refactor: change program structure
