+++ src/Dto/User/AuthUserDto.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace ECommerce.src.Dto
+namespace ECommerce.src.Dto.User
 {
     public class AuthUserDto
     {
