+++ src/Dto/User/RegisterDto.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 
 
-namespace ECommerce.src.Dto
+namespace ECommerce.src.Dto.User
 {
     public class RegisterDto
     {
