+++ src/Mappers/UserMapper.cs
@@ -4,7 +4,7 @@ using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 
-using ECommerce.src.Dto;
+using ECommerce.src.Dto.User;
 using ECommerce.src.Models;
 
 using Microsoft.AspNetCore.StaticAssets;
